requirements.txt
feedparser
pyyaml
bs4
feedgen
requests
pytz
html2text